// widget/src/types/theme.ts
export interface WidgetTheme {
  cardBg: string;
  border: string;

  heading: string;
  title: string;
  author: string;
  pages: string;

  progressBar: string;
  progressTrack: string;
  percentage: string;

  cardWidth: number;
  cardHeight: number;

  borderWidth: number;
  borderRadius: number;
}